+++ examples/service/verify/webhook/fetch/webhook_fetch_example.go
@@ -24,19 +24,14 @@ func init() {
 }
 
 func main() {
-	paginator := verifySession.
+	resp, err := verifySession.
 		Service("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
-		RateLimits.
-		NewRateLimitsPaginator()
+		Webhook("YWXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		Fetch()
 
-	for paginator.Next() {
-		currentPage := paginator.CurrentPage()
-		log.Printf("%v rate limit(s) found on page %v", len(currentPage.RateLimits), currentPage.Meta.Page)
+	if err != nil {
+		log.Panicf("%s", err.Error())
 	}
 
-	if paginator.Error() != nil {
-		log.Panicf("%s", paginator.Error())
-	}
-
-	log.Printf("Total number of rate limit(s) found: %v", len(paginator.RateLimits))
+	log.Printf("SID: %s", resp.Sid)
 }